feat: convert to typescript
* Converts module to typescript
* Upgrades all deps
* Makes sure we're actually running the CI jobs we think we are

Also fixes a bug with the cache TTL, previously we were doing:

```js
Math.min(response.Answer.map(a => a.TTL))
```

which returns `NaN`.  Changes it to:

```js
Math.min(...response.Answer.map(a => a.TTL))
```

which returns a number.

BREAKING CHANGE: ESM only publishing
