+++ .github/workflows/main.yml
@@ -15,7 +15,7 @@ jobs:
     - run: npm install
     - run: npm run lint
     - run: npm run build
-    - run: npx aegir dep-check
+    - run: npm run dep-check
     - uses: ipfs/aegir/actions/bundle-size@master
       name: size
       with:
@@ -26,7 +26,7 @@ jobs:
     strategy:
       matrix:
         os: [windows-latest, ubuntu-latest, macos-latest]
-        node: [14, 15]
+        node: [16]
       fail-fast: true
     steps:
       - uses: actions/checkout@v2
@@ -34,7 +34,7 @@ jobs:
         with:
           node-version: ${{ matrix.node }}
       - run: npm install
-      - run: npx nyc --reporter=lcov aegir test -t node -- --bail
+      - run: npx nyc --reporter=lcov npm run test -- -t node -- --bail
       - uses: codecov/codecov-action@v1
   test-chrome:
     needs: check
@@ -42,11 +42,11 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - run: npm install
-      - run: npx aegir test -t browser --bail
+      - run: npm run test -- -t browser --bail
   test-firefox:
     needs: check
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
       - run: npm install
-      - run: npx aegir test -t browser --bail -- --browsers FirefoxHeadless
+      - run: npm run test -- -t browser --bail -- --browser firefox
