+++ package.json
@@ -2,25 +2,50 @@
   "name": "dns-over-http-resolver",
   "version": "1.2.3",
   "description": "DNS over HTTP resolver",
-  "main": "src/index.js",
   "author": "Vasco Santos",
-  "scripts": {
-    "test": "aegir test -t node -t browser",
-    "test:browser": "aegir test -t browser",
-    "test:node": "aegir test -t node",
-    "lint": "aegir lint",
-    "release": "aegir release --docs",
-    "release-minor": "aegir release --type minor --docs",
-    "release-major": "aegir release -t node -t browser --type major --docs",
-    "build": "aegir build"
-  },
+  "type": "module",
   "files": [
     "src",
-    "dist"
+    "dist/src"
   ],
+  "exports": {
+    ".": {
+      "import": "./dist/src/index.js",
+      "types": "./dist/src/index.d.ts"
+    }
+  },
+  "typesVersions": {
+    "*": {
+      "src/*": [
+        "dist/src/*",
+        "dist/src/*/index"
+      ]
+    }
+  },
+  "eslintConfig": {
+    "extends": "ipfs",
+    "parserOptions": {
+      "sourceType": "module"
+    }
+  },
+  "scripts": {
+    "pretest": "npm run build",
+    "test": "aegir test -f ./dist/test/*.js",
+    "test:browser": "aegir test -t browser -f ./dist/test/*.js",
+    "test:node": "aegir test -t node -f ./dist/test/*.js",
+    "test:electron": "aegir test -t electron-main -f ./dist/test/*.js",
+    "lint": "aegir ts -p check && aegir lint",
+    "release": "aegir release --docs",
+    "release-minor": "aegir release --target node --type minor --docs",
+    "release-major": "aegir release --type major --docs",
+    "build": "tsc",
+    "dep-check": "aegir dep-check dist/src/**/*.js dist/test/**/*.js"
+  },
   "devDependencies": {
-    "aegir": "^33.1.0",
-    "sinon": "^11.1.2",
+    "@types/debug": "^4.1.7",
+    "aegir": "^36.0.2",
+    "sinon": "^12.0.1",
+    "ts-sinon": "^2.0.2",
     "util": "^0.12.3"
   },
   "repository": {
@@ -39,7 +64,7 @@
   "homepage": "https://github.com/vasco-santos/dns-over-http-resolver#readme",
   "dependencies": {
     "debug": "^4.3.1",
-    "native-fetch": "^3.0.0",
+    "native-fetch": "^4.0.2",
     "receptacle": "^1.3.2"
   },
   "contributors": [
